DAOTHER-8066 enabled chunk forwarding, rewrote locks, addressed race conditions, fixed move file test

+++ rcsb/app/file/uploadRequest.py
@@ -4,35 +4,28 @@
 #
 ##
 __docformat__ = "google en"
-__author__ = "John Westbrook"
-__email__ = "john.westbrook@rcsb.org"
+__author__ = "James Smith"
+__email__ = "james.smith@rcsb.org"
 __license__ = "Apache 2.0"
 
 import gzip
 import hashlib
-import io
 import logging
 import os
 import re
 import shutil
 import uuid
-import json
 from enum import Enum
 from typing import Optional
-from urllib.parse import unquote
-
+from filelock import Timeout, FileLock
 import aiofiles
-import fastapi
-import fastapi.responses
-from fastapi.responses import PlainTextResponse
-from fastapi import APIRouter, Path, Query, Request, Depends, File, Form, HTTPException, UploadFile, Body
+from fastapi import APIRouter, Path, Query, File, Form, HTTPException, UploadFile
 from pydantic import BaseModel  # pylint: disable=no-name-in-module
 from pydantic import Field
 from rcsb.app.file.ConfigProvider import ConfigProvider
 from rcsb.app.file.IoUtils import IoUtils
-from rcsb.app.file.JWTAuthBearer import JWTAuthBearer
 from rcsb.app.file.PathUtils import PathUtils
-from rcsb.utils.io.FileLock import FileLock
+# from rcsb.utils.io.FileLock import FileLock
 
 
 logger = logging.getLogger(__name__)
@@ -72,7 +65,6 @@ async def upload(
     uploadId: str = Form(None),
     hashType: HashType = Form(None),
     hashDigest: str = Form(None),
-    copyMode: str = Form("native"),
     # save file parameters
     depId: str = Form(...),
     repositoryType: str = Form(...),
@@ -81,6 +73,7 @@ async def upload(
     partNumber: int = Form(...),
     contentFormat: str = Form(...),
     version: str = Form(None),
+    copyMode: str = Form("native"),
     allowOverwrite: bool = Form(None)
 ):
     tempPath = None
@@ -90,35 +83,34 @@ async def upload(
     try:
         fn = uploadFile.filename
         ct = uploadFile.content_type
-        # if fn.endswith(".gz") or ct == "application/gzip":
-        #     copyMode = "gzip_decompress"
         cachePath = os.environ.get("CACHE_PATH")
         configFilePath = os.environ.get("CONFIG_FILE")
         cP = ConfigProvider(cachePath, configFilePath)
         pathU = PathUtils(cP)
         if not pathU.checkContentTypeFormat(contentType, contentFormat):
             raise HTTPException(status_code=405, detail = "Bad content type and/or format - upload rejected")
-        lockPath = pathU.getFileLockPath(depId, contentType, milestone, partNumber, contentFormat)
-        with FileLock(lockPath):
-            outPath = pathU.getVersionedPath(
-                repositoryType=repositoryType,
-                depId=depId,
-                contentType=contentType,
-                milestone=milestone,
-                partNumber=partNumber,
-                contentFormat=contentFormat,
-                version=version
-            )
+        outPath = pathU.getVersionedPath(
+            repositoryType=repositoryType,
+            depId=depId,
+            contentType=contentType,
+            milestone=milestone,
+            partNumber=partNumber,
+            contentFormat=contentFormat,
+            version=version
+        )
         if not outPath:
             raise HTTPException(status_code=405, detail = "Bad content type metadata - cannot build a valid path")
         if os.path.exists(outPath) and not allowOverwrite:
             raise HTTPException(status_code=405, detail = "Encountered existing file - overwrite prohibited")
         dirPath, fn = os.path.split(outPath)
-        tempPath = os.path.join(dirPath, "." + fn)
+        uploadId = await getNewUploadId()
+        uploadId = uploadId["id"]
+        tempPath = os.path.join(dirPath, "." + uploadId)
         os.makedirs(dirPath, mode=0o755, exist_ok=True)
-        if copyMode == "shell" or copyMode == "native" or copyMode == "gzip_decompress":
-            with open(tempPath, "wb") as ofh:
-                shutil.copyfileobj(ifh, ofh)
+        # save (all copy modes), then hash, then decompress
+        with open(tempPath, "wb") as ofh:
+            shutil.copyfileobj(ifh, ofh)
+        # hash
         if hashDigest and hashType:
             if hashType == "SHA1":
                 hashObj = hashlib.sha1()
@@ -135,23 +127,40 @@ async def upload(
             if not ok:
                 raise HTTPException(status_code=400, detail = f"{hashType} hash check failed {hexdigest} != {hashDigest}")
             else:
-                os.replace(tempPath, outPath)
+                # lock before saving
+                lockPath = os.path.join(os.path.dirname(outPath), "." + os.path.basename(outPath) + ".lock")
+                lock = FileLock(lockPath)
+                try:
+                    with lock.acquire(timeout = 60 * 60 * 4):
+                        # last minute race condition prevention
+                        if os.path.exists(outPath) and not allowOverwrite:
+                            raise HTTPException(status_code=400, detail="Encountered existing file - cannot overwrite")
+                        else:
+                            # save final version
+                            os.replace(tempPath, outPath)
+                except Timeout:
+                    raise HTTPException(status_code=400, detail=f"error - file lock timed out on {outPath}")
+                finally:
+                    lock.release()
+                    if os.path.exists(lockPath):
+                        os.unlink(lockPath)
+                # decompress
                 if copyMode == "gzip_decompress":
                     with gzip.open(outPath, "rb") as r:
                         with open(tempPath, "wb") as w:
                             w.write(r.read())
                     os.replace(tempPath, outPath)
-            if os.path.exists(tempPath):
-                os.unlink(tempPath)
         else:
             raise HTTPException(status_code=500, detail = "Error - missing hash")
-    except HTTPException as e:
-        ret = {"success": False, "statusCode": e.status_code, "statusMessage": f"Store fails with {e.detail}"}
+    except HTTPException as exc:
+        ret = {"success": False, "statusCode": exc.status_code, "statusMessage": f"Store fails with {exc.detail}"}
         logging.warning(ret["statusMessage"])
-    except Exception as e:
-        ret = {"success": False, "statusCode": 400, "statusMessage": f"Store fails with {str(e)}"}
+    except Exception as exc:
+        ret = {"success": False, "statusCode": 400, "statusMessage": f"Store fails with {str(exc)}"}
         logging.warning(ret["statusMessage"])
     finally:
+        if tempPath and os.path.exists(tempPath):
+            os.unlink(tempPath)
         ifh.close()
     if not ret["success"]:
         raise HTTPException(status_code=405, detail=ret["statusMessage"])
@@ -179,18 +188,16 @@ async def getSaveFilePath(repositoryType: str = Query(...),
     pathU = PathUtils(cP)
     if not pathU.checkContentTypeFormat(contentType, contentFormat):
         raise HTTPException(status_code=400, detail="Bad content type and/or format - upload rejected")
-    lockPath = pathU.getFileLockPath(depId, contentType, milestone, partNumber, contentFormat)
     outPath = None
-    with FileLock(lockPath):
-        outPath = pathU.getVersionedPath(
-            repositoryType=repositoryType,
-            depId=depId,
-            contentType=contentType,
-            milestone=milestone,
-            partNumber=partNumber,
-            contentFormat=contentFormat,
-            version=version
-        )
+    outPath = pathU.getVersionedPath(
+        repositoryType=repositoryType,
+        depId=depId,
+        contentType=contentType,
+        milestone=milestone,
+        partNumber=partNumber,
+        contentFormat=contentFormat,
+        version=version
+    )
     if not outPath:
         raise HTTPException(status_code=400, detail="Bad content type metadata - cannot build a valid path")
     if os.path.exists(outPath) and not allowOverwrite:
@@ -198,7 +205,6 @@ async def getSaveFilePath(repositoryType: str = Query(...),
     dirPath, fn = os.path.split(outPath)
     os.makedirs(dirPath, mode=0o755, exist_ok=True)
     return {"path": outPath}
-    #return PlainTextResponse(outPath)
 
 # upload chunk
 @router.post("/sequentialUpload")
@@ -208,7 +214,6 @@ async def sequentialUpload(
     uploadId: str = Form(None),
     hashType: str = Form(None),
     hashDigest: str = Form(None),
-    copyMode: str = Form("native"),
     # chunk parameters
     chunkSize: int = Form(None),
     chunkIndex: int = Form(None),
@@ -216,54 +221,89 @@ async def sequentialUpload(
     expectedChunks: int = Form(None),
     chunkMode: str = Form("sequential"),
     # save file parameters
-    filePath: str = Form(...)
+    filePath: str = Form(...),
+    copyMode: str = Form("native"),
+    allowOverwrite: bool = Query(default=False)
 ):
-    contents = await uploadFile.read()
-    if len(contents) <= 0:
-        raise HTTPException(status_code=400, detail="error - empty file")
     ret = {"success": True, "statusCode": 200, "statusMessage": "Chunk uploaded"}
     fn = uploadFile.filename
     ct = uploadFile.content_type
-    # if fn.endswith(".gz") or ct == "application/gzip":
-    #    copyMode = "gzip_decompress"
     dirPath, fn = os.path.split(filePath)
     tempPath = os.path.join(dirPath, "." + uploadId)
-    async with aiofiles.open(tempPath, "ab") as ofh:
-        await ofh.seek(chunkOffset)
-        await ofh.write(contents)
-        await ofh.flush()
-        os.fsync(ofh.fileno())
-    if chunkIndex + 1 == expectedChunks:
-        if hashDigest and hashType:
-            if hashType == "SHA1":
-                hashObj = hashlib.sha1()
-            elif hashType == "SHA256":
-                hashObj = hashlib.sha256()
-            elif hashType == "MD5":
-                hashObj = hashlib.md5()
-            blockSize = 65536
-            with open(tempPath, "rb") as r:
-                while chunk := r.read(blockSize):
-                    hashObj.update(chunk)
-            hexdigest = hashObj.hexdigest()
-            ok = (hexdigest == hashDigest)
-            if not ok:
-                raise HTTPException(status_code=400, detail=f"{hashType} hash check failed")
+    contents = await uploadFile.read()
+    # empty chunk beyond loop index from client side, don't erase tempPath so keep out of try block
+    if contents and len(contents) <= 0:
+        raise HTTPException(status_code=400, detail="error - empty file")
+    try:
+        # save, then hash, then decompress
+        # should lock, however client must wait for each response before sending next chunk, precluding race conditions (unless multifile upload problem)
+        async with aiofiles.open(tempPath, "ab") as ofh:
+            await ofh.seek(chunkOffset)
+            await ofh.write(contents)
+            await ofh.flush()
+            os.fsync(ofh.fileno())
+        # if last chunk
+        if chunkIndex + 1 == expectedChunks:
+            if hashDigest and hashType:
+                if hashType == "SHA1":
+                    hashObj = hashlib.sha1()
+                elif hashType == "SHA256":
+                    hashObj = hashlib.sha256()
+                elif hashType == "MD5":
+                    hashObj = hashlib.md5()
+                blockSize = 65536
+                # hash
+                with open(tempPath, "rb") as r:
+                    while chunk := r.read(blockSize):
+                        hashObj.update(chunk)
+                hexdigest = hashObj.hexdigest()
+                ok = (hexdigest == hashDigest)
+                if not ok:
+                    raise HTTPException(status_code=400, detail=f"{hashType} hash check failed")
+                else:
+                    # lock then save
+                    lockPath = os.path.join(os.path.dirname(filePath), "." + os.path.basename(filePath) + ".lock")
+                    lock = FileLock(lockPath)
+                    try:
+                        with lock.acquire(timeout = 60 * 60 * 4):
+                            # last minute race condition handling
+                            if os.path.exists(filePath) and not allowOverwrite:
+                                raise HTTPException(status_code=400, detail="Encountered existing file - cannot overwrite")
+                            else:
+                                # save final version
+                                os.replace(tempPath, filePath)
+                    except Timeout:
+                        raise HTTPException(status_code=400, detail=f'error - lock timed out on {filePath}')
+                    finally:
+                        lock.release()
+                        if os.path.exists(lockPath):
+                            os.unlink(lockPath)
+                # remove temp file
+                if os.path.exists(tempPath):
+                    os.unlink(tempPath)
+                # decompress
+                if copyMode == "gzip_decompress":
+                    with gzip.open(filePath, "rb") as r:
+                        with open(tempPath, "wb") as w:
+                            w.write(r.read())
+                    os.replace(tempPath, filePath)
             else:
-                os.replace(tempPath, filePath)
-            if os.path.exists(tempPath):
-                os.unlink(tempPath)
-            if copyMode == "gzip_decompress":
-                with gzip.open(filePath, "rb") as r:
-                    with open(tempPath, "wb") as w:
-                        w.write(r.read())
-                os.replace(tempPath, filePath)
-        else:
-            raise HTTPException(status_code=500, detail="Error - missing hash")
+                raise HTTPException(status_code=500, detail="Error - missing hash")
+    except HTTPException as exc:
+        if os.path.exists(tempPath):
+            os.unlink(tempPath)
+        ret = {"success": False, "statusCode": exc.status_code, "statusMessage": f"error in sequential upload {exc.detail}"}
+        raise HTTPException(status_code=exc.status_code, detail=exc.detail)
+    except Exception as exc:
+        if os.path.exists(tempPath):
+            os.unlink(tempPath)
+        ret = {"success": False, "statusCode": 400, "statusMessage": f"error in sequential upload {str(exc)}"}
+        raise HTTPException(status_code=400, detail=f'error in sequential upload {str(exc)}')
     return ret
 
 
-# return kv entry from file parameters, or None
+# return kv entry from file parameters, if have resumed upload, or None if don't
+# if have resumed upload, kv response has chunk indices and count
 @router.get("/uploadStatus")
 async def getUploadStatus(repositoryType: str = Query(...),
                           depId: str = Query(...),
@@ -315,14 +355,14 @@ async def getUploadStatusFromId(uploadId: str = Path(...)):
 
 
 # find upload id from file parameters
-@router.post("/findUploadId")
-async def findUploadId(repositoryType: str = Form(),
-                       depId: str = Form(...),
-                       contentType: str = Form(...),
-                       milestone: str = Form(...),
-                       partNumber: int = Form(...),
-                       contentFormat: str = Form(...),
-                       version: str = Form(...)):
+@router.get("/findUploadId")
+async def findUploadId(repositoryType: str = Query(),
+                       depId: str = Query(...),
+                       contentType: str = Query(...),
+                       milestone: str = Query(...),
+                       partNumber: int = Query(...),
+                       contentFormat: str = Query(...),
+                       version: str = Query(...)):
     cachePath = os.environ.get("CACHE_PATH")
     configFilePath = os.environ.get("CONFIG_FILE")
     cP = ConfigProvider(cachePath, configFilePath)
@@ -357,13 +397,12 @@ async def asyncUpload(
     uploadId: str = Form(None),
     hashType: HashType = Form(None),
     hashDigest: str = Form(None),
-    copyMode: str = Form("native"),
     # chunk parameters
     chunkSize: int = Form(None),
     chunkIndex: int = Form(0),
     chunkOffset: int = Form(0),
     expectedChunks: int = Form(1),
-    chunkMode: str = Form("sequential"),
+    chunkMode: str = Form("async"),
     # save file parameters
     depId: str = Form(...),
     repositoryType: str = Form(...),
@@ -372,6 +411,7 @@ async def asyncUpload(
     partNumber: int = Form(...),
     contentFormat: str = Form(...),
     version: str = Form(...),
+    copyMode: str = Form("native"),
     allowOverwrite: bool = Form(None),
     emailAddress: str = Form(None)
 ):
@@ -385,11 +425,9 @@ async def asyncUpload(
         fn = uploadFile.filename
         ct = uploadFile.content_type
         logger.debug("uploadFile %s (%r)", fn, ct)
-        # if fn.endswith(".gz") or ct == "application/gzip":
-        #     copyMode = "gzip_decompress"
         logger.debug("hashType.name %r hashDigest %r", hashType, hashDigest)
         ioU = IoUtils(cP)
-        ret = await ioU.storeUpload(
+        ret = await ioU.asyncUpload(
             # upload file parameters
             ifh=uploadFile.file,
             uploadId=uploadId,
@@ -397,6 +435,7 @@ async def asyncUpload(
             hashDigest=hashDigest,
             copyMode=copyMode,
             # chunk parameters
+            chunkSize=chunkSize,
             chunkIndex=chunkIndex,
             chunkOffset=chunkOffset,
             expectedChunks=expectedChunks,
